backend/pyproject.toml
[tool.poetry]
name = "KEEN.uz Backend"
version = "0.1.0"
description = ""
authors = ["Nazarbek", "Olimboy"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django-class-settings = "*"
bs4 = "*"
daphne = "*"
django = "4.2.9"
django-ckeditor = "*"
django-cors-headers = "*"
django-extensions = "*"
django-filter = "*"
django-modeltranslation = "*"
django-nested-inline = "*"
django-q = "*"
django-quill-editor = "*"
django-redis-cache = "*"
djangorestframework = "*"
djangorestframework-camel-case = "*"
social-auth-app-django = "^5.3.0"
django-storages = "^1.14.1"
django-solo = "^2.1.0"
psycopg2-binary = "^2.9.9"
psycopg2 = "^2.9.9"
pillow = "*"
boto3 = "^1.28.59"
psutil = "^5.9.5"
drf-extra-fields = "^3.7.0"
whitenoise = "^6.5.0"
redis = "^5.0.1"
websockets = "^11.0.3"
websocket-client = "^1.6.3"
django-hint = "^0.3.2"
openpyxl = "^3.1.2"
lxml = "^4.9.3"
javalang = "^0.13.0"
gunicorn = "^21.2.0"
drf-yasg = "*"
channels = "^4.0.0"
channels-redis = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
